Refactor and security improve users controllers

+++ src/server/controllers/notes.controllers.js
@@ -5,36 +5,41 @@ import Note from "../models/Note.model.js";
 import Concept from "../models/Concept.model.js";
 import Language_Concept from "../models/Language_Concept.model.js";
 import Category from "../models/Category.model.js";
-import { Op } from "sequelize";
+import { Validator } from "../utils/Validator.js";
+
+//¨ Controllers
 
 //* @ DELETE /api/notes/delete/:pk
 //* @ Delete a concept with primaryKey
 async function DeleteNote(req, res) {
-  //~ Delete note with it primaryKey
   let { pk } = req.params;
-  let index = await Note.destroy({ where: { id_note: pk } });
 
-  //~ Send response to the client
-  index
-    ? res
-      .status(200)
-      .send({ index: `Note '${pk}' has been delete succesfully !` })
-    : res.status(404).send({ index: `Note '${pk}' hasn't been find !` });
+  // Control primaryKey
+  if (!Validator.num(pk)) return res.status(400).send({ error: `The primaryKey '${pk}' isn't a number !` })
+
+  // Delete note with it primaryKey
+  let result = await Note.destroy({ where: { id_note: pk } });
+
+  // Send response to the client
+  if (result) res.status(200).send({ index: `Note '${pk}' has been delete succesfully !` })
+  else res.status(404).send({ index: `Note '${pk}' hasn't been find !` });
 }
 
 //* @ PUT /api/notes/update/
 //* @ Update a concept
 async function UpdateConcept(req, res) {
-  //~ Find and update note with its primaryKey
   let { pk, percentage } = req.body;
-  let index = await Note.update({ percentage }, { where: { id_note: pk } });
 
-  //~ Send response to the client
-  index[0]
-    ? res
-      .status(200)
-      .send({ index: `Note '${pk}' has been update succesfully !` })
-    : res.status(404).send({ index: `Note '${pk}' hasn't been find !` });
+  // Control primaryKey and percentage
+  if (!Validator.num(pk)) return res.status(400).send({ error: `The primaryKey '${pk}' isn't a number !` })
+  if (!Validator.percentage(percentage)) return res.status(400).send({ error: `The percentage '${percentage}' isn't valid !` })
+
+  // Find and update note with its primaryKey
+  let result = await Note.update({ percentage }, { where: { id_note: pk } });
+
+  // Send response to the client
+  if (result[0]) res.status(200).send({ result: `Note '${pk}' has been update succesfully !` })
+  else res.status(404).send({ error: `Note '${pk}' hasn't been find !` });
 }
 
 //* @ GET /api/notes/
@@ -42,43 +47,35 @@ async function UpdateConcept(req, res) {
 async function GetAllCategories(req, res) {
   let categories = await Category.findAll({ raw: true })
 
-  //~ Create a map object
+  // Create a map object
   let notes = new Map();
   categories.forEach((v) => { notes.set(v.id_category, { name: v.name, note: [] }) });
 
-  //~ Get all the languages by category, return languages primary_Key and id cat
-  let languages = await Language.findAll({
-    attributes: ["id_language", "id_category"],
-    raw: true,
-  });
+  // Get all the languages by category, return languages primary_Key and id cat
+  let languages = await Language.findAll({ attributes: ["id_language", "id_category"], raw: true, });
 
-  //~ Fill the object with id-cat as key and the associated notes
+  // Fill the object with id-cat as key and the associated notes
   for (const l of languages) {
-    let id_category = l.id_category;
+    let { id_category, id_language } = l;
 
-    //~ Find the value in the object having the index as Id_cat
+    // Find the value in the object having the index as Id_cat
     let index = notes.get(id_category);
 
-    //~ Get the last upadted note for each language
-    let note = await Note.findOne({
-      where: { id_language: l.id_language },
-      attributes: ["percentage"],
-      order: [["date", "DESC"]],
-      raw: true,
-    });
+    // Get the last upadted note for each language
+    let note = await Note.findOne({ where: { id_language }, attributes: ["percentage"], order: [["date", "DESC"]], raw: true, });
 
-    //~ Regroup notes by category
+    // Regroup notes by category
     if (index && note) {
       index.note.push(note.percentage);
       notes.delete(id_category);
       notes.set(id_category, index);
-    } //~ If id_cat doesnt match an index, a new index is created with it's note
+    } // If id_cat doesnt match an index, a new index is created with it's note
     else if (note) notes.set(id_category, [note.percentage]);
   }
 
 
   let result = [];
-  //~ Calcul the average value of each category
+  // Calcul the average value of each category
   for (let [k, v] of notes) {
     let average = 0;
     for (const n of v.note) n ? average += n : 0;
@@ -98,7 +95,7 @@ async function GetAllCategories(req, res) {
 
   result = Array.from(notes.values())
 
-  //~ Send the response to the client
+  // Send the response to the client
   result
     ? res.status(200).send({ result })
     : res.status(404).send({ result: `Data hasn't been find !` });
@@ -106,18 +103,15 @@ async function GetAllCategories(req, res) {
 
 //* @ GET /api/notes/category/:pk
 //* @ Get languages and their notes by category
+//TODO Refactor maybe
 async function GetLanguagesOfCategory(req, res) {
-  //~ Get all the languages with id_category
+  // Get all the languages with id_category
   let { id_category } = req.params;
-  let languages = await Language.findAll({
-    where: { id_category },
-    attributes: { exclude: ["id_category"] },
-    raw: true,
-  });
+  let languages = await Language.findAll({ where: { id_category }, attributes: { exclude: ["id_category"] }, raw: true });
 
   let result = [];
 
-  // //~ Get notes, dates for each languages push them in an array [key=>value]
+  // // Get notes, dates for each languages push them in an array [key=>value]
   for (const v of languages) {
     let { id_language } = v
     let request = await Note.findAll({
@@ -145,7 +139,7 @@ async function GetLanguagesOfCategory(req, res) {
     result.push({ name: v.name, notes, dates, id: id_language })
   }
 
-  //~ Send the response to the client
+  // Send the response to the client
   result
     ? res.status(200).send({ result })
     : res.status(404).send({ result: `Data hasn't been find !` });
@@ -153,8 +147,9 @@ async function GetLanguagesOfCategory(req, res) {
 
 //* @ GET /api/notes/language/:pk
 //* @ Get each concepts with its value for a language
+//TODO Refactor maybe
 async function GetConceptsOfLanguage(req, res) {
-  //~ Get all id_concept with its id_language
+  // Get all id_concept with its id_language
   let { id_language } = req.params;
   let id_concepts = await Language_Concept.findAll({
     where: { id_language },
@@ -162,7 +157,7 @@ async function GetConceptsOfLanguage(req, res) {
     attributes: ["id_concept"],
   });
 
-  //~ Get name and value for each concept with its primaryKey
+  // Get name and value for each concept with its primaryKey
   let result = [];
   for (const v of id_concepts) {
     let concept = await Concept.findByPk(v.id_concept, {
@@ -172,7 +167,7 @@ async function GetConceptsOfLanguage(req, res) {
     result.push(concept);
   }
 
-  //~ Send response to the client
+  // Send response to the client
   result[0]
     ? res.status(200).send({ result })
     : res.status(404).send({
@@ -199,7 +194,7 @@ function parseDate(date) {
   return d
 }
 
-//~Export all the function created in this present file to whatever other file would ask for.
+//Export all the function created in this present file to whatever other file would ask for.
 export {
   DeleteNote,
   UpdateConcept,
